Fixed it :)
Switcvhed to the arduino ide and now coding in C using various setup tools we have the accelerometer data working! The z direction is a bit messed up (offset by a fixed amount) but it is accurate. Gyro also works. Just need to figure out how to export serial data for other use